+++ lib/private/CommitQuery.d.ts
@@ -1,0 +1,4 @@
+import { Query } from '../Query';
+export declare class CommitQuery extends Query {
+    getQuery(): string;
+}