+++ lib/private/RollbackQuery.d.ts
@@ -1,0 +1,4 @@
+import { Query } from '../Query';
+export declare class RollbackQuery extends Query {
+    getQuery(): string;
+}