+++ lib/private/StartTransactionQuery.d.ts
@@ -1,0 +1,4 @@
+import { Query } from '../Query';
+export declare class StartTransactionQuery extends Query {
+    getQuery(): string;
+}