+++ src/private/CommitQuery.ts
@@ -1,0 +1,8 @@
+
+import {Query} from '../Query';
+
+export class CommitQuery extends Query {
+    public getQuery(): string {
+        return 'COMMIT';
+    }
+}