+++ src/private/RollbackQuery.ts
@@ -1,0 +1,8 @@
+
+import {Query} from '../Query';
+
+export class RollbackQuery extends Query {
+    public getQuery(): string {
+        return 'ROLLBACK';
+    }
+}