+++ src/private/StartTransactionQuery.ts
@@ -1,0 +1,8 @@
+
+import {Query} from '../Query';
+
+export class StartTransactionQuery extends Query {
+    public getQuery(): string {
+        return 'START TRANSACTION';
+    }
+}